Add files via upload
This is the script that combines the CSV files from our web scrapes of the fake company generator and outputs a single data frame. From there, the script removes stop words, removes apostrophes, performs tokenization and lemmatization then assigns sentiment scores to the purpose of each company. Finally, we sort by the sentiment score and output the top and bottom 5 companies by the score of their purpose